// src/app/lib/generated/models/ChatImageCategory.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ChatImageCategory = {
    Welcome: 'welcome',
    ExposedAss: 'exposed_ass',
    ExposedBreasts: 'exposed_breasts',
    ExposedPussy: 'exposed_pussy',
    Sex: 'sex',
    Toys: 'toys'
} as const;
export type ChatImageCategory = typeof ChatImageCategory[keyof typeof ChatImageCategory];


export function instanceOfChatImageCategory(value: any): boolean {
    for (const key in ChatImageCategory) {
        if (Object.prototype.hasOwnProperty.call(ChatImageCategory, key)) {
            if (ChatImageCategory[key as keyof typeof ChatImageCategory] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ChatImageCategoryFromJSON(json: any): ChatImageCategory {
    return ChatImageCategoryFromJSONTyped(json, false);
}

export function ChatImageCategoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChatImageCategory {
    return json as ChatImageCategory;
}

export function ChatImageCategoryToJSON(value?: ChatImageCategory | null): any {
    return value as any;
}

export function ChatImageCategoryToJSONTyped(value: any, ignoreDiscriminator: boolean): ChatImageCategory {
    return value as ChatImageCategory;
}