// src/app/lib/generated/models/PictureModel.ts
export interface PictureModel {
    /**
     * 
     * @type {Date}
     * @memberof PictureModel
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PictureModel
     */
    updatedAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof PictureModel
     */
    id?: string;
    /**
     * 
     * @type {ContentType}
     * @memberof PictureModel
     */
    type: ContentType;
    /**
     * 
     * @type {ChatImageCategory}
     * @memberof PictureModel
     */
    chatCategory: ChatImageCategory | null;
    /**
     * 
     * @type {string}
     * @memberof PictureModel
     */
    postId: string | null;
    /**
     * 
     * @type {string}
     * @memberof PictureModel
     */
    galleryId: string | null;
    /**
     * 
     * @type {string}
     * @memberof PictureModel
     */
    chatbotId: string;
    /**
     * 
     * @type {string}
     * @memberof PictureModel
     */
    cloudFilename: string;
    /**
     * 
     * @type {string}
     * @memberof PictureModel
     */
    prompt: string;
    /**
     * 
     * @type {string}
     * @memberof PictureModel
     */
    description: string | null;
}